MB-32245: Fix shared_ptr race in spdlog registry
In changing the spdlog registry to use weak_ptr's, a race condition
was introduced. A shared_ptr to a BucketLogger may be destructed,
dropping the internal reference count to 0. If any other thread blocks
this thread from unregistering the BucketLogger in the BucketLogger
destructor by acquiring the logger_map_mutex_, then the blocking thread
can access the loggers_ map in an invalid state that can cause a
segmentation fault. This is best described by a diagram found in the
comments on the jira issue.

To fix this, check the validity of the shared_ptr we create when we
call lock() on the stored weak_ptr. If it's not valid, ignore the entry
in the map. The thread currently blocked by the mutex in the drop
function (called by unregisterSpdlogger) will cleanup the map. If we
were to cleanup the map when we find an invalid weak_ptr, we would open
ourselves to another potential issue where we could register another
logger with the same name on a different thread, then unregister it
immediately with the destructing thread (if you ignore that we try to
guarantee unique names). This wouldn't seg fault, but we wouldn't be
able to change the verbosity of the newly created logger.

Change-Id: I54f475901ff26940ecedb1cabc37a29e4347a46c